Add more guide attributes.
But not all of the ones available to us, because some will need more advanced
data types that I don't want to deal with yet.

Also, remove the static list of attributes, because there are a lot of them and
I don't want to maintain two lists.  dir(guide) will just have to suffer.