Add Guide.all() to fetch lists of Guides.
I'm not adding tests yet because I know any that test guideids-specifying will
fail; the API doesn't take into account offset or limit when given guideids,
and I'm not sure yet if that's a bug or a failure of the documentation.

Is that a good excuse for not writing the tests?  No, not really.  I admit it -
the real reason is because I'm anxious to use this, and I've now completed
enough functionality for the project that spawned this.  Go ahead, call me a
terrible TDDer - I just don't particularly care at the moment.