setup.py: gracefully handle lack of pandoc
We were catching the module not existing or the file not existing, but failed
to account for the `pandoc` binary missing from the system.  Surprise!  Pandoc
isn't installed on Read the Docs, so it started failing builds.

RtD was previously successful because we didn't include pypandoc in
`requirements.txt`, so it failed earlier on with the `ImportError`.